[Fix] fixture 삭제
테스트를 위한 Request는 Fixture 보다는 Mock에 가깝다고 판단
