+++ src/test/java/codesquad/filter/MethodOverrideFilterTest.java
@@ -1,6 +1,9 @@
 package codesquad.filter;
 
 import codesquad.fixture.http.MethodFieldRequestFixture;
+import codesquad.http.MockRequest;
+import codesquad.http.MockRequestDispatcher;
+import codesquad.http.MockResponse;
 import jakarta.servlet.ServletException;
 import org.junit.jupiter.api.*;
 
@@ -9,11 +12,16 @@ import java.io.IOException;
 class MethodOverrideFilterTest implements MethodFieldRequestFixture {
     private MethodOverrideFilter filter;
     private MockFilterChain filterChain;
+    private MockRequest mockRequest;
+    private MockResponse mockResponse;
 
     @BeforeEach
     void setUp() {
         filter = new MethodOverrideFilter();
         filterChain = new MockFilterChain();
+        MockRequestDispatcher mockRequestDispatcher = new MockRequestDispatcher();
+        mockRequest = new MockRequest("/something", "POST", mockRequestDispatcher);
+        mockResponse = new MockResponse();
     }
 
     @Nested
@@ -22,7 +30,7 @@ class MethodOverrideFilterTest implements MethodFieldRequestFixture {
         @Test
         @DisplayName("그대로 POST으로 처리된다.")
         void processAsPost() throws ServletException, IOException {
-            filter.doFilter(postRequest(), emptyResponse(), filterChain);
+            filter.doFilter(postFieldRequest(mockRequest), mockResponse, filterChain);
 
             Assertions.assertEquals(filterChain.getRequest().getMethod(), "POST");
         }
@@ -30,7 +38,7 @@ class MethodOverrideFilterTest implements MethodFieldRequestFixture {
         @Test
         @DisplayName("_method 필드가 PUT이면 PUT으로 처리된다.")
         void processAsPut() throws ServletException, IOException {
-            filter.doFilter(putFieldRequest(), emptyResponse(), filterChain);
+            filter.doFilter(putFieldRequest(mockRequest), mockResponse, filterChain);
 
             Assertions.assertEquals(filterChain.getRequest().getMethod(), "PUT");
         }
@@ -38,7 +46,7 @@ class MethodOverrideFilterTest implements MethodFieldRequestFixture {
         @Test
         @DisplayName("_method 필드가 DELETE이면 DELETE으로 처리된다.")
         void processAsDelete() throws ServletException, IOException {
-            filter.doFilter(deleteFieldRequest(), emptyResponse(), filterChain);
+            filter.doFilter(deleteFieldRequest(mockRequest), mockResponse, filterChain);
 
             Assertions.assertEquals(filterChain.getRequest().getMethod(), "DELETE");
         }
